+++ public/index.html
@@ -10,7 +10,7 @@
       <button class="log" onclick="log()">LOG</button>
     </header>
     <main>
-      <div style="margin: 20vh 0 4vh">
+      <div style="margin: 15vh 0 4vh">
         <time id="time">33:33</time>
       </div>
       <div id="line">
